added admin check, on refresh on submitting forms

+++ routes/web.php
@@ -5,6 +5,7 @@ use App\Models\Round;
 use App\Models\Competition;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UserController;
+use App\Http\Middleware\AdminMiddleware;
 use App\Http\Controllers\RoundController;
 use App\Http\Controllers\CompetitorController;
 use App\Http\Controllers\CompetitionController;
@@ -15,12 +16,12 @@ Route::get('/', function () {
 
 Route::get('/addComp', function () {
     return view('addComp');
-});
+})->middleware(AdminMiddleware::class);
 
 Route::get('/addRound', function () {
     $competitions = Competition::all();
     return view('addRound', compact('competitions'));
-});
+})->middleware(AdminMiddleware::class);
 
 Route::get('/addCompetitor', function () {
     $competitions = Competition::all();
@@ -29,7 +30,7 @@ Route::get('/addCompetitor', function () {
     return view('addCompetitor', compact('competitions', 'users', 'rounds'));
 });
 
-Route::delete('/deleteCompetitors/{id}', [CompetitorController::class, 'delete']);
+Route::delete('/deleteCompetitors/{id}', [CompetitorController::class, 'delete'])->middleware(AdminMiddleware::class);
 
 Route::get('/competitors', [CompetitorController::class, 'index']);
